[VOGRE-15] Added project stat helper function to utils, comments
